Create plot for machine learning
